Homework#8 room db

+++ app/src/main/java/ru/suslovalex/androidacademyapp/data/ActorResponse.kt
@@ -1,11 +1,10 @@
 package ru.suslovalex.androidacademyapp.data
 
 
-import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class ActorResponse(
-    val id: Int,
+    val id: Long,
     val cast: List<Cast>
 )
